Changing default OS user password.
Setting the default user password to "Changeme" for new box installs for ease of setup.
Adding a randomly-generated user password as a vault variable in group_vars. This allows for rotating the password in a central way.
Setting the default user's password to the random value. Keys are required to SSH into the machine, and with the random password, users won't be able to su into the default account either.